+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Change `script run hf_bruteforce -s start_id -e end_id -t timeout -x mifare_card_type` - brute force Mifare card script now requires type (mfc or mfu) (@dunderhay)
  - Updated `hf_bruteforce.lua` script - added support for brute forcing Mifare Ultralight EV1 cards (@dunderhay)
  - Added `hf mf personlize` - personalize the UID of a Mifare Classic EV1 card (@pwpiwi)
  - Change - hint texts added to all lf clone commands (@iceman1001)